requirements.txt
opencv-python==4.7.0.68
mediapipe==0.9.0.1
scikit-learn==1.2.0

numpy~=1.24.4
pyttsx3~=2.90
pillow~=10.2.0